Fix bugs about buf socket.
In previous commit, two bugs related to buf socket.The first is the
epoll flags, it use logic or instead of logic and, causing the entire
pseudo-mm-rdma-server busy in epoll loop and waste cpu cycle. The second
is fill the memory pool with memory image. In CRIU, after sending the fd
of image, the CRIU converter process might exit, causing the client side
of buf socket closed. After that pseudo-mm-rdma-server cannot read from
socket, intead it rece a HUP epoll event and miss thoses memory image
fds. The solution need to modify criu and rdma-server both, after CRIU
sends the fds, it need to recv an ack from server. And the server will
send an ack after fill the memory pool with memory images.